% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_data_chatgpt.R
\name{analyze_data_chatgpt}
\alias{analyze_data_chatgpt}
\title{Title}
\usage{
analyze_data_chatgpt(
  vector,
  temperature = 0.3,
  openai_api_key,
  model = "gpt-3.5-turbo",
  questions = "question 1: summarise this:",
  sleep = 5
)
}
\arguments{
\item{sleep}{}
}
\description{
Title
}
