% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comments_data.R
\name{get_comments_data}
\alias{get_comments_data}
\title{Title}
\usage{
get_comments_data(
  subreddit,
  ncomments = 10,
  max_error_conseq = 3,
  grouped = FALSE,
  subset = TRUE,
  subsetn = 13000
)
}
\arguments{
\item{subsetn}{}
}
\description{
Title
}
