% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_full_dataset.R
\name{get_full_dataset}
\alias{get_full_dataset}
\title{Title}
\usage{
get_full_dataset(
  subreddit = "",
  nsubmissions = 10,
  ncomments = 10,
  max_error_conseq = 3,
  subset = TRUE,
  subsetn = 13000
)
}
\arguments{
\item{subsetn}{}
}
\description{
Title
}
