% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_submission_data.R
\name{get_submission_data}
\alias{get_submission_data}
\title{Title}
\usage{
get_submission_data(subreddit, nsubmissions = 10, max_error_conseq = 3)
}
\arguments{
\item{max_error_conseq}{}
}
\description{
Title
}
